+++ mlx/semseg/camvid.py
@@ -14,7 +14,7 @@ from fastai.callbacks import SaveModelCallback, CSVLogger, Callback
 
 import mlx.semseg.s3_utils as s3_utils
 import mlx.semseg.batch_utils as batch_utils
-from mlx.semseg.utils import make_dir
+from mlx.utils import make_dir
 
 
 class S3SyncCallback(Callback):
